GRC_Agent-main/requirements.txt
streamlit
langchain
python-dotenv
huggingface_hub>=0.30.0
sentence-transformers
langchain-community
langchain==0.3.27
langchain-core==0.3.72
langchain-community>=0.0.22
langchain-google-genai>=0.0.10
google-search-results==2.4.2
python-dotenv==1.1.1
nltk
google.generativeai
arabic_reshaper
python-bidi
langchain_huggingface
transformers
einops
hf-transfer
certifi
langgraph
IPython
qdrant_client
langchain_qdrant
langchain_huggingface
pydantic
tqdm
ragas
pandas
numexpr
langchainhub
sentencepiece 
jinja2
bitsandbytes
accelerate
rapidfuzz
langchain-deepseek
langchain-huggingface
langchain_core
dotenv 